Fixed errors

+++ src/main/java/pl/wwsis/sos/dao/KsiazkaDao.java
@@ -5,8 +5,8 @@ import pl.wwsis.sos.model.Ksiazka;
 import java.util.List;
 
 public interface KsiazkaDao {
-    Ksiazka findById(Long id);
+    Ksiazka findById(Integer id);
     List<Ksiazka> findAll();
     void save(Ksiazka ksiazka);
-    void delete(Long id);
+    void delete(Integer id);
 }
